+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - [#38](https://github.com/Cloud-Harvest/CloudHarvestCoreTasks/issues/38): Address `'NoneType' object has no attriute 'append'` in the `TaskError` exception class.
  - `BaseTask` now offers the `prefix` property used for log entries.
 - `HarvestUpdateTask.bulk_replace()` now chunks large replacements to prevent overwhelming the database.
+- Added `filter_datetime_ago()` jinja2 filter.
 
 ## 0.8.1
 - Removed `meta.json` in favor of using `pyproject.toml`