+++ CHANGELOG.md
@@ -5,7 +5,9 @@
 - [#37](https://github.com/Cloud-Harvest/CloudHarvestCoreTasks/issues/37): Log messages now include the calling filename, line number, and task name. 
 - [#38](https://github.com/Cloud-Harvest/CloudHarvestCoreTasks/issues/38): Address `'NoneType' object has no attriute 'append'` in the `TaskError` exception class.
  - `BaseTask` now offers the `prefix` property used for log entries.
-- `HarvestUpdateTask.bulk_replace()` now chunks large replacements to prevent overwhelming the database.
+- `HarvestUpdateTask.bulk_replace()`
+  - now chunks large replacements to prevent overwhelming the database.
+  - now retries writes as sometimes the backend database is too busy to complete the request at that time.
 - Added `filter_datetime_ago()` jinja2 filter.
 
 ## 0.8.1